Adjustement for tox 4
Tox version 4 no more supports the `[tox:jenkins]` section. That was
used to load a different flask configuration file when running under
Jenkins.

Move the config file selection to the application and set the
configuration file when the `CI` environment variable is set. Can be
tried locally with:

  CI=1 tox run -e test

Remove the unused `__dir__` variable in tests_main.py which confuses
nose2 when it iterates the module attributes. It is not used anyway.

Tox 4 changed the way it processes comments which causes inline comments
in a multiline value to not be passed as is. That broke flake8 `ignore`
option.

Since flake8 `ignore` option overrides the default built-in value:
- switch to use `extend_ignore` to append to the default list
- remove W504 (line break after binary operator) which is ignored by
  default).

Finally remove the comment indentation to let tox recognize it as a
comment.

Run DB operations under an application context.

Bug: T345607
Change-Id: I7585afc5d3200f2a86a79ac9dd2174b535e4401f